src/requirements.txt
djangorestframework==3.12.2
Django==3.1.7
selenium==3.141.0
gunicorn==20.0.4
selenium==3.141.0
PyVirtualDisplay==2.0